[BP-1] - refactoring
