+++ src/main/kotlin/com/bp/app/account/domain/AccountMapper.kt
@@ -1,0 +1,10 @@
+package com.bp.app.account.domain
+
+import com.bp.app.account.service.AccountDto
+
+fun Account.toDto() = AccountDto(
+    id = id!!,
+    name = name!!,
+    phone = phone!!,
+    email = email!!
+)
