+++ src/main/kotlin/com/bp/app/account/service/AccountDto.kt
@@ -1,4 +1,4 @@
-package com.bp.app.account
+package com.bp.app.account.service
 
 data class AccountDto(
     val id: Long,
