+++ src/main/kotlin/com/bp/app/account/service/AccountService.kt
@@ -1,9 +1,9 @@
 package com.bp.app.account.service
 
-import com.bp.app.account.Account
-import com.bp.app.account.AccountDto
-import com.bp.app.account.AccountRepository
+import com.bp.app.account.domain.AccountRepository
+import com.bp.app.account.domain.toDto
 import com.bp.app.account.web.request.SignUpRequest
+import com.bp.app.account.web.request.toEntity
 import org.springframework.stereotype.Service
 
 @Service
@@ -13,17 +13,3 @@ class AccountService(private val repository: AccountRepository) {
         return repository.save(request.toEntity()).toDto()
     }
 }
-
-fun SignUpRequest.toEntity() = Account(
-    name = name,
-    phone = phone,
-    email = email,
-    password = password
-)
-
-fun Account.toDto() = AccountDto(
-    id = id!!,
-    name = name!!,
-    phone = phone!!,
-    email = email!!
-)
